+++ packages/esm-amd-loader/test/static/cycle/b.js
@@ -1,0 +1,7 @@
+define(['exports', './a.js'], function(exports, a) {
+  exports.b = 'b';
+  exports.usesAAtExecution = a.a;
+  exports.getterForA = function() {
+    return a.a;
+  }
+});