+++ packages/esm-amd-loader/test/static/deepRace/b.js
@@ -1,0 +1,3 @@
+define(['./c.js', './d.js'], function() {
+  window.executionOrder.push('b');
+});