+++ packages/esm-amd-loader/test/static/deepRace/start-two.js
@@ -1,0 +1,3 @@
+define(['./a.js', './g.js', './h.js'], function() {
+  window.executionOrder.push('start-two');
+});