+++ packages/esm-amd-loader/test/static/failure/failure.js
@@ -1,0 +1,4 @@
+define([], function() {
+  window.executionOrder.push('failure');
+  throw new Error('failure.js is supposed to fail');
+});